Moved ActiveUsersHeaderDropdown from Whiteboard.tsx to its own component file.
