+++ Frontend/package.json
@@ -31,6 +31,7 @@
     "dotenv": "^17.2.3",
     "its-fine": "^2.0.0",
     "konva": "^9.3.22",
+    "lodash": "^4.18.1",
     "lucide-react": "^0.546.0",
     "path": "^0.12.7",
     "radix-ui": "^1.4.3",
@@ -50,6 +51,7 @@
   "devDependencies": {
     "@eslint/js": "^9.30.1",
     "@types/axios": "^0.14.4",
+    "@types/lodash": "^4.17.24",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "@vitejs/plugin-react-swc": "^3.10.2",
